system: convert bad to std::vector

+++ include/slvs.h
@@ -151,11 +151,11 @@ typedef struct {
      * param[] is actually an in/out variable; if the solver is successful,
      * then the new values (that satisfy the constraints) are written to it. */
     Slvs_Param          *param;
-    int                 params;
+    size_t              params;
     Slvs_Entity         *entity;
-    int                 entities;
+    size_t              entities;
     Slvs_Constraint     *constraint;
-    int                 constraints;
+    size_t              constraints;
 
     /* If a parameter corresponds to a point (distance, normal, etc.) being
      * dragged, then specify it here. This will cause the solver to favor
@@ -184,7 +184,7 @@ typedef struct {
      * failing constraints, faileds should be greater than or equal to
      * constraints. */
     Slvs_hConstraint    *failed;
-    int                 faileds;
+    size_t              faileds;
 
     /* The solver indicates the number of unconstrained degrees of freedom. */
     int                 dof;
@@ -202,7 +202,7 @@ typedef struct {
     int                 result;
     int                 dof;
     int                 rank;
-    int                 bad;
+    size_t              bad;
 } Slvs_SolveResult;
 
 /* Our base coordinate system has basis vectors
